Move more stuff out of index.ts
