+++ js/hang/src/publish/index.ts
@@ -1,10 +1,10 @@
-export * from "./audio";
+export * as Audio from "./audio";
 export * from "./broadcast";
 export * from "./chat";
 export * from "./location";
 export * from "./preview";
 export * as Source from "./source";
-export * from "./video";
+export * as Video from "./video";
 
 // NOTE: element is not exported from this module
 // You have to import it from @kixelated/hang/publish/element instead.
